+++ src/ui/components/Banner.tsx
@@ -11,6 +11,7 @@ import { DogLogo } from "./DogLogo.js";
 import { KunLogo } from "./KunLogo.js";
 import { KunHead } from "./KunHead.js";
 import { BabyHand } from "./BabyHand.js";
+import { getCurrentVersion } from "../../cli/updater.js";
 
 // Logo selection: random from all four (or CODESHELL_LOGO env to force one)
 type LogoKind = "dog" | "kunhead" | "kunball" | "baby";
@@ -57,7 +58,7 @@ export function Banner({ model, effort, cwd }: BannerProps) {
         <Box flexDirection="column">
           <Box>
             <Text bold>{title}</Text>
-            <Text dim>{" v0.1.0"}</Text>
+            <Text dim>{" v" + getCurrentVersion()}</Text>
           </Box>
           <Box>
             <Text>{modelName}</Text>